pyproject.toml
[project]
name = "bl531"
version = "0.1.0"
description = "bl531 - Osprey Agent Application"
requires-python = ">=3.11"

dependencies = [
    # Framework dependency
    "osprey-framework>=0.8.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "ruff",
]

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["bl531*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

